Add read checkbox and moved delete event listeners
The event listener function for the delete buttons was creating extra
listeners every time it was called instead of being overwritten in the
var deleteButtons like I expected. This approach became a problem when
implementing a similar function for the read checkboxes. The event
listeners are now attached to the element as soon as it is created in
the displayBooksOnPage function.